requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dj-database-url==2.3.0
dj-rest-auth==7.0.0
Django==5.1.4
django-allauth==65.3.0
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
h11==0.14.0
idna==3.10
inflection==0.5.1
packaging==24.2
psycopg2==2.9.10
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
whitenoise==6.8.2
